requirements.txt
python-dotenv
chainlit
agentic